fix the course home page and the course view page

+++ courses/src/main/java/com/knet51/courses/jpa/services/CommentService.java
@@ -9,7 +9,7 @@ import com.knet51.ccweb.jpa.entities.teacher.Comment;
 public interface CommentService {
 	Comment createComment(Comment comment);
 	Page<Comment> findAllCommit(int pageNumber, int pageSize,Long teacherCourse_id);
-	Long getMark(Long teacherCourseId);
+	Double getMark(Long teacherCourseId);
 	Long getPerson(Long teacherCourseId);
 	Comment getComment(Long teacherCourseId, Long userId);
 	int  getCommentByTeacherCourseIdAndUserId(Long teacherCourseId, Long userId);
