+++ data-access-layer/src/main/java/com/knet51/ccweb/jpa/repository/CommentRepositoryCustom.java
@@ -8,7 +8,7 @@ import com.knet51.ccweb.jpa.entities.teacher.Comment;
 public interface CommentRepositoryCustom {
 	Comment createComment(Comment comment);
 	Long getPerson(Long  teacherCourseId);
-	Long getMark(Long teacherCourseId);
+	Double getMark(Long teacherCourseId);
 	Comment getComment(Long teacherCourseId, Long userId);
 	int  getCommentByTeacherCourseIdAndUserId(Long teacherCourseId, Long userId);
 	List<Comment> getAllCourse(Long  teacherCourseId);
